+++ Tests/TABResourceLoaderTests/Protocols/NetworkResourceTypeTests.swift
@@ -21,6 +21,7 @@ private struct MockCustomNetworkResource: NetworkResourceType {
   let httpHeaderFields: [String: String]?
   let jsonBody: Any?
   let queryItems: [URLQueryItem]?
+  let requestTimeoutInterval: TimeInterval? = 27
 
   init(url: URL, httpRequestMethod: HTTPMethod = .get, httpHeaderFields: [String : String]? = nil, jsonBody: Any? = nil, queryItems: [URLQueryItem]? = nil) {
     self.url = url
@@ -35,13 +36,21 @@ class NetworkResourceTypeTests: XCTestCase {
 
   let url = URL(string: "www.test.com")!
   let urlWithQueryItem = URL(string: "www.test.com?query-name=query-value")!
+  // using the defaut time interval dynamically instead of hard-coding a value that may change at
+  // some point of time
+  private lazy var defaultRequestTimeOut: TimeInterval = {
+    let request = URLRequest(url: url)
+    return request.timeoutInterval
+  }()
 
   func test_correctDefaultValues() {
     let resource = MockDefaultNetworkResource(url: url)
+    let request = resource.urlRequest()
     XCTAssertEqual(resource.httpRequestMethod, HTTPMethod.get)
     XCTAssertEqual(resource.httpHeaderFields!, [:])
     XCTAssertNil(resource.jsonBody)
     XCTAssertNil(resource.queryItems)
+    XCTAssertEqual(request?.timeoutInterval, defaultRequestTimeOut)
   }
 
   func test_urlRequest_allProperties() {
@@ -59,6 +68,7 @@ class NetworkResourceTypeTests: XCTestCase {
     XCTAssertEqual(urlRequest!.allHTTPHeaderFields!, expectedAllHTTPHeaderFields)
     let expectedJSONData = try? JSONSerialization.data(withJSONObject: expectedJSONBody, options: JSONSerialization.WritingOptions.prettyPrinted)
     XCTAssertEqual(urlRequest!.httpBody, expectedJSONData)
+    XCTAssertEqual(urlRequest?.timeoutInterval, 27)
   }
 
   func test_urlRequest_resourceURLWithQueryParameters() {